fix to allow control connection to properly report failed connections if msgtype is invalid.
also fix of cntrl->state to actually notice that test sessions are over.

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@370 a2d82226-602c-0410-bba7-e51b83dd10d2

+++ owamp/owampP.h
@@ -344,7 +344,7 @@ _OWPReadServerOK(
 	OWPAcceptType	*acceptval	/* ret	*/
 	);
 
-extern u_int8_t
+extern int
 OWPReadRequestType(
 	OWPControl	cntrl
 	);
